requirements.txt
# Smart Job Hunter — Python dependencies
# Install with: pip install -r requirements.txt
# Then run:     playwright install chromium

# LangChain stack
langchain
langchain-openai
langchain-core

# Browser automation
playwright==1.56.0

# Environment variables
python-dotenv==1.2.1

# PDF extraction
pdfplumber

# Utils
pydantic
openai

# Testing
pytest